+++ NEWS.md
@@ -3,6 +3,7 @@
 * left to right assignment is now supported in side effects
 * New pipes `%P.%`, `%<P.%`, `%V.%`,  and `%<V.%`, are introduced to `print()` or
 `View()` the result of every step.
+* We can use `if` and `else` to have conditional steps
 
 # nakedpipe 0.0.0.9000